+++ ranger-core/src/main/java/io/appform/ranger/core/finderhub/ServiceFinderHub.java
@@ -26,6 +26,7 @@ import io.appform.ranger.core.signals.ExternalTriggeredSignal;
 import io.appform.ranger.core.signals.ScheduledSignal;
 import io.appform.ranger.core.signals.Signal;
 import io.appform.ranger.core.util.Exceptions;
+import io.appform.ranger.core.util.FinderUtils;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
@@ -52,7 +53,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
             new AtomicReference<>(new ConcurrentHashMap<>());
     private final Lock updateLock = new ReentrantLock();
     private final Condition updateCond = updateLock.newCondition();
-    private boolean updateAvailable = false;
+    private final AtomicBoolean updateAvailable = new AtomicBoolean(false);
     private final ExecutorService executorService = Executors.newFixedThreadPool(1);
 
     @Getter
@@ -72,8 +73,10 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
     private final AtomicInteger poolThreadIndex = new AtomicInteger(0);
     private Future<?> monitorFuture = null;
 
-    private final long serviceRefreshDurationMs;
-    private final long hubRefreshDurationMs;
+    private final long serviceRefreshTimeoutMs;
+    private final long hubStartTimeoutMs;
+
+    private final Set<String> excludedServices;
 
     private final ForkJoinPool refresherPool;
 
@@ -82,23 +85,25 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
             ServiceFinderFactory<T, R> finderFactory
                            ) {
         this(serviceDataSource, finderFactory,
-             HubConstants.SERVICE_REFRESH_DURATION_MS, HubConstants.HUB_REFRESH_DURATION_MS);
+                HubConstants.SERVICE_REFRESH_TIMEOUT_MS, HubConstants.HUB_START_TIMEOUT_MS, Set.of());
     }
 
     public ServiceFinderHub(
             ServiceDataSource serviceDataSource,
             ServiceFinderFactory<T, R> finderFactory,
-            long serviceRefreshDurationMs,
-            long hubRefreshDurationMs) {
+            long serviceRefreshTimeoutMs,
+            long hubStartTimeoutMs,
+            final Set<String> excludedServices) {
         this.serviceDataSource = serviceDataSource;
         this.finderFactory = finderFactory;
-        this.serviceRefreshDurationMs = serviceRefreshDurationMs == 0 ? HubConstants.SERVICE_REFRESH_DURATION_MS : serviceRefreshDurationMs;
-        this.hubRefreshDurationMs = hubRefreshDurationMs == 0 ? HubConstants.HUB_REFRESH_DURATION_MS : hubRefreshDurationMs;
+        this.serviceRefreshTimeoutMs = serviceRefreshTimeoutMs == 0 ? HubConstants.SERVICE_REFRESH_TIMEOUT_MS : serviceRefreshTimeoutMs;
+        this.hubStartTimeoutMs = hubStartTimeoutMs == 0 ? HubConstants.HUB_START_TIMEOUT_MS : hubStartTimeoutMs;
         this.refreshSignals.add(new ScheduledSignal<>("service-hub-updater",
                                                       () -> null,
                                                       Collections.emptyList(),
-                                                      10_000));
+                                      10_000));
         this.refresherPool = createRefresherPool();
+        this.excludedServices = Objects.requireNonNullElseGet(excludedServices, Set::of);
     }
 
     public Optional<ServiceFinder<T, R>> finder(final Service service) {
@@ -155,7 +160,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
     public void updateAvailable() {
         try {
             updateLock.lock();
-            updateAvailable = true;
+            updateAvailable.set(true);
             updateCond.signalAll();
         }
         finally {
@@ -179,7 +184,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
         while (true) {
             try {
                 updateLock.lock();
-                while (!updateAvailable) {
+                while (!updateAvailable.get()) {
                     updateCond.await();
                 }
                 updateRegistry();
@@ -190,7 +195,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
                 break;
             }
             finally {
-                updateAvailable = false;
+                updateAvailable.set(false);
                 updateLock.unlock();
             }
         }
@@ -204,7 +209,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
         alreadyUpdating.set(true);
         val updatedFinders = new ConcurrentHashMap<Service, ServiceFinder<T, R>>();
         try {
-            val services = serviceDataSource.services();
+            val services = FinderUtils.getEligibleServices(serviceDataSource.services(), excludedServices);
             if (services.isEmpty()) {
                 log.debug("No services found for the service data source. Skipping update on the registry");
                 return;
@@ -239,13 +244,13 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
     }
 
     private void waitTillHubIsReady() {
-        val services = serviceDataSource.services();
-        val timeToRefresh = Math.max(hubRefreshDurationMs,
-                                     (serviceRefreshDurationMs * services.size()) / refresherPool.getParallelism());
-        if (timeToRefresh != hubRefreshDurationMs) {
+        val services = FinderUtils.getEligibleServices(serviceDataSource.services(), excludedServices);
+        val timeToRefresh = Math.max(hubStartTimeoutMs,
+                                     (serviceRefreshTimeoutMs * services.size()) / refresherPool.getParallelism());
+        if (timeToRefresh != hubStartTimeoutMs) {
             log.warn("Max hub refresh time has been dynamically adjusted to {} ms from the provided {} ms as the " +
                              "provided time would have been insufficient to refresh {} services.",
-                     timeToRefresh, hubRefreshDurationMs, services.size());
+                    timeToRefresh, hubStartTimeoutMs, services.size());
         }
         val hubRefresher = CompletableFuture.allOf(
                 services.stream()
@@ -273,7 +278,7 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
         try {
             RetryerBuilder.<Boolean>newBuilder()
                     .retryIfResult(r -> !r)
-                    .withStopStrategy(StopStrategies.stopAfterDelay(serviceRefreshDurationMs, TimeUnit.MILLISECONDS))
+                    .withStopStrategy(StopStrategies.stopAfterDelay(serviceRefreshTimeoutMs, TimeUnit.MILLISECONDS))
                     .build()
                     .call(() -> Optional.ofNullable(getFinders().get().get(service))
                             .map(ServiceFinder::getServiceRegistry)
@@ -285,4 +290,5 @@ public class ServiceFinderHub<T, R extends ServiceRegistry<T>> {
                     .illegalState("Could not perform initial state for service: " + service.getServiceName(), e);
         }
     }
+
 }