+++ ranger-http-client/src/main/java/io/appform/ranger/client/http/AbstractRangerHttpHubClient.java
@@ -63,8 +63,9 @@ public abstract class AbstractRangerHttpHubClient<T, R extends ServiceRegistry<T
                 .withServiceDataSource(getServiceDataSource())
                 .withServiceFinderFactory(getFinderFactory())
                 .withRefreshFrequencyMs(getNodeRefreshTimeMs())
-                .withHubRefreshDuration(getHubRefreshDurationMs())
-                .withServiceRefreshDuration(getServiceRefreshDurationMs())
+                .withHubStartTimeout(getHubStartTimeoutMs())
+                .withServiceRefreshTimeout(getServiceRefreshTimeoutMs())
+                .withExcludedServices(getExcludedServices())
                 .build();
     }
 }