+++ ranger-zk-client/src/main/java/io/appform/ranger/client/zk/AbstractRangerZKHubClient.java
@@ -46,8 +46,9 @@ public abstract class AbstractRangerZKHubClient<T, R extends ServiceRegistry<T>,
                 .withRefreshFrequencyMs(getNodeRefreshTimeMs())
                 .withServiceDataSource(getServiceDataSource())
                 .withServiceFinderFactory(getFinderFactory())
-                .withHubRefreshDuration(getHubRefreshDurationMs())
-                .withServiceRefreshDuration(getServiceRefreshDurationMs())
+                .withHubStartTimeout(getHubStartTimeoutMs())
+                .withServiceRefreshTimeout(getServiceRefreshTimeoutMs())
+                .withExcludedServices(getExcludedServices())
                 .build();
     }